Move contstruct data
